// src/interfaces/gestion-producto/superlinea/interfaces-superlinea.tsx
export interface Superlinea {
  id: number;
  denominacion: string;
  observacion: string;
  createdAt: string | null;
  updatedAt: string | null;
  deletedAt: string | null;
  usuarioCreatedId?: number;
  usuarioDeletedId?: number;
  usuarioUpdatedId?: number;
  sistema: number;
}

export interface SelectSuperlinea {
  id: number;
  denominacion: string;
}

export interface DtoConsultarSuperlinea {
  data: Superlinea[];
  total: number;
}